+++ workers/ohsome_quality_analyst/api/api.py
@@ -24,6 +24,7 @@ from ohsome_quality_analyst.api.request_models import (
 )
 from ohsome_quality_analyst.geodatabase import client as db_client
 from ohsome_quality_analyst.utils.definitions import (
+    INDICATOR_LAYER,
     configure_logging,
     get_dataset_names,
     get_fid_fields,
@@ -238,6 +239,14 @@ async def get_available_regions(asGeoJSON: bool = False):
     else:
         regions = await db_client.get_regions()
         response["result"] = regions
+    return response
+
+
+@app.get("/indicatorLayerCombinations")
+async def list_indicator_layer_combinations():
+    """List names of available indicator-layer-combinations."""
+    response = empty_api_response()
+    response["result"] = INDICATOR_LAYER
     return response