+++ workers/tests/integrationtests/test_api.py
@@ -41,6 +41,13 @@ class TestApi(unittest.TestCase):
         self.assertTrue(self.general_schema.is_valid(response_content))
         self.assertIsInstance(response_content["result"], list)
 
+    def test_list_indicator_layer_combinations(self):
+        url = "/indicatorLayerCombinations"
+        response = self.client.get(url)
+        self.assertEqual(response.status_code, 200)
+
+        self.assertIsInstance(response.json(), dict)
+
     def test_list_indicators(self):
         url = "/indicatorNames"
         response = self.client.get(url)