+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Add and improve tests for API ([#168])
     - Add test cases covering the content of GeoJSON properties.
     - Improve API response schemata by using less logic to create the schemata.
+- Add list indicator/layer combinations for API and CLI([#99])
 - Load indicator from DB will also load its data attributes ([#179])
 - Added parameter to regions endpoint to only respond with a geojson if requested. ([#171])
 
@@ -26,6 +27,7 @@
 
 - If you set up your own database you will need to rebuild the database or delete the results table (`DROP TABLE results;`).
 
+[#99]: https://github.com/GIScience/ohsome-quality-analyst/issues/99
 [#102]: https://github.com/GIScience/ohsome-quality-analyst/pull/102
 [#106]: https://github.com/GIScience/ohsome-quality-analyst/issues/106
 [#141]: https://github.com/GIScience/ohsome-quality-analyst/pull/141